[cmake][mbrhack] Drop dependency on the parted tool
This change removes from Tilck entirely the dependency on the `parted` tool
by using `mbrhack` instead. Now mbrhack can be used to write a standard MBR
to the given image and the creation of partitions is the same as before.
We're using MBR for both the legacy/hybrid boot mode and for the pure UEFI
boot mode as well, as it officially supported by the UEFI standard itself,
even if it's technically not the best practice. In the future, we'll use GPT
for the pure UEFI boot, of course.